Fixed warning in OsFileIteratorBase
git-svn-id: https://sipxsvn.sipez.com/rep/sipX/main@12660 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXportLib/include/os/OsFileIteratorBase.h
@@ -1,6 +1,5 @@
 //
-// Copyright (C) 2005 SIPez LLC.
-// Licensed to SIPfoundry under a Contributor Agreement.
+// Copyright (C) 2005-2017 SIPez LLC.  All rights reserved.
 //
 // Copyright (C) 2004-2006 SIPfoundry Inc.
 // Licensed by SIPfoundry under the LGPL license.
@@ -101,7 +100,7 @@ protected:
     RegEx*      mFilterExp;
       //: The regular expression that the user searched for
 
-    long mSearchHandle;
+    intptr_t mSearchHandle;
 
     // release memory that allocated for mFilterExp 
     //
